Fix repositoires imports

+++ FCT/src/main/java/IFC33B/FCT/repository/TutorEmpresaRepository.java
@@ -1,9 +1,11 @@
 package IFC33B.FCT.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
-import IFC33B.FCT.model.Tutor;
+import IFC33B.FCT.model.TutorEmpresa;
 
 public interface TutorEmpresaRepository extends JpaRepository<TutorEmpresa, Long> {
-    Optional<TutorEmpresa> findById(Integer Id);
-    boolean existsById(Integer Id);
+    Optional<TutorEmpresa> findById(Long Id);
+    boolean existsById(Long Id);
 }
